fix(ci): python-full-coverage — cancel stale monitor recovery before handoff finalize and split oversized test modules
When stale-monitor reconcile cancels a resume task before the remonitor op is finalized, classify the operation as cancelled (not failed) and drop the recovery handle so the claim finally block cannot overwrite it; split worker part 047 and coverage-edges part 001 to satisfy the 1500-line maintainability gate.
