+++ tests/unit/control/test_worker_coverage_edges_parts/test_worker_coverage_edges_part_001.py
@@ -43,7 +43,6 @@ from awf.db.repositories import (
     SecretLeaseRepository,
     TaskAttemptRepository,
     TaskRepository,
-    WorkspaceEventRepository,
     WorkspaceRepository,
 )
 from awf.db.resilience import (
@@ -1426,75 +1425,6 @@ async def test_stale_active_execution_can_fail_rejects_preserved_runtime(
         await session.commit()
 
     assert not await worker._stale_active_execution_can_fail(  # noqa: SLF001
-        _ActiveExecutionCandidate(
-            workspace_id=workspace_id,
-            status=WorkspaceStatus.running,
-            compose_project_name=f"awf_{workspace_id}",
-        )
-    )
-
-
-@pytest.mark.unit
-async def test_stale_active_execution_can_fail_ignores_salvage_for_other_status(
-    factory: async_sessionmaker[AsyncSession],
-) -> None:
-    workspace_id = await _seed_status(
-        factory, WorkspaceStatus.running, title="status-scoped-salvage"
-    )
-    now = datetime.now(UTC)
-    status_started_at = now - timedelta(minutes=10)
-    claim_expires_at = now - timedelta(minutes=5)
-    preserved_at = now - timedelta(minutes=4)
-    mismatched_salvage_at = now - timedelta(minutes=3)
-    stale_at = now - timedelta(minutes=2)
-
-    async with factory() as session:
-        repo = WorkspaceRepository(session)
-        ws = await repo.get(workspace_id)
-        assert ws is not None
-        ws.execution_claimed_by = "stale-worker"
-        ws.execution_claim_expires_at = claim_expires_at
-        state_events = await WorkspaceEventRepository(session).list(
-            workspace_id=workspace_id,
-            event_type="workspace.state_changed",
-        )
-        running_started = next(
-            event for event in state_events if event.new_state == WorkspaceStatus.running.value
-        )
-        running_started.occurred_at = status_started_at
-        preserved = await repo.add_event(
-            ws,
-            event_type=ACTIVE_EXECUTION_PRESERVED_EVENT_TYPE,
-            reason_code=ACTIVE_EXECUTION_PRESERVED_REASON_CODE,
-            payload={"workspace_status": WorkspaceStatus.running.value},
-        )
-        preserved.occurred_at = preserved_at
-        salvage = await repo.add_event(
-            ws,
-            event_type="workspace.active_execution_salvage_operator_required",
-            reason_code="ACTIVE_EXECUTION_SALVAGE_OPERATOR_REQUIRED",
-            payload={
-                "reason_code": "ACTIVE_EXECUTION_SALVAGE_OPERATOR_REQUIRED",
-                "workspace_status": WorkspaceStatus.validating.value,
-            },
-        )
-        salvage.occurred_at = mismatched_salvage_at
-        stale = await repo.add_event(
-            ws,
-            event_type=_STALE_ACTIVE_EXECUTION_EVENT_TYPE,
-            reason_code=_STALE_ACTIVE_EXECUTION_REASON_CODE,
-            payload={"workspace_status": WorkspaceStatus.running.value},
-        )
-        stale.occurred_at = stale_at
-        await session.commit()
-
-    worker = ControlWorker(
-        session_factory=factory,
-        provisioner=_NoopProvisioner(),  # type: ignore[arg-type]
-        config=WorkerConfig(active_execution_preservation_grace_seconds=0.0),
-    )
-
-    assert await worker._stale_active_execution_can_fail(  # noqa: SLF001
         _ActiveExecutionCandidate(
             workspace_id=workspace_id,
             status=WorkspaceStatus.running,
